Reset function
If N is left for three seconds then the user jumps back to the scoring stage. If button A or B is pressed and Y is selected and left for three seconds all the saved files are erased and the vars are reset.

The user is then taken back to the setTeams stage.